Refactor: vcompute move vld and vst
